+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v3.70.2-4] - 2024-10-30
+### Fixed
+- [#148] Fix writing the claim once lock into wrong key
+
 ## [v3.70.2-3] - 2024-09-25
 ### Fixed
 - Local config was created by root with doguctl-call during pre-startup in multinode.